Move loadSupplementalFHIRPackage to load.ts
